consolidated code of database services
