+++ byoeb-v1/byoeb/byoeb/background_jobs/app_config.json
@@ -21,7 +21,8 @@
             "provider_type": "azure_cosmos_mongo_db",
             "database_name": "ashadb",
             "user_collection": "ashausers",
-            "message_collection": "ashamessages"
+            "message_collection": "ashamessages",
+            "jobs_collection": "ashajobs"
         }
     },
     "llms": {
