+++ byoeb-v1/byoeb/byoeb/background_jobs/message_leaderboard/leaderboard.py
@@ -6,41 +6,10 @@ import pandas as pd
 
 from byoeb.background_jobs.config import app_config
 from byoeb.chat_app.configuration import dependency_setup
-from byoeb.services.leaderboard import LeaderboardService
+from byoeb.background_jobs.dependency_setup import get_leaderboard_service, get_user_service, get_message_service
 from byoeb.services.leaderboard.time_window_strategies import TimeWindowFactory
-from byoeb.services.user import UserService
-from byoeb.services.message import MessageService
 
 IST = ZoneInfo("Asia/Kolkata")
-
-# Service instances
-_leaderboard_service: Optional[LeaderboardService] = None
-_user_service: Optional[UserService] = None
-_message_service: Optional[MessageService] = None
-
-async def get_leaderboard_service() -> LeaderboardService:
-    """Get or create leaderboard service instance."""
-    global _leaderboard_service
-    if _leaderboard_service is None:
-        user_service = await get_user_service()
-        message_service = await get_message_service()
-        _leaderboard_service = LeaderboardService(user_service, message_service)
-    return _leaderboard_service
-
-async def get_user_service() -> UserService:
-    """Get or create user service instance."""
-    global _user_service
-    if _user_service is None:
-        _user_service = UserService()
-    return _user_service
-
-async def get_message_service() -> MessageService:
-    """Get or create message service instance."""
-    global _message_service
-    if _message_service is None:
-        user_service = await get_user_service()
-        _message_service = MessageService(user_service)
-    return _message_service
 
 async def fetch_phone_numbers_for_asha_and_test_users() -> List[str]:
     """
@@ -64,7 +33,9 @@ async def build_district_leaderboard_last_week_ist(message_categories: Optional[
         pd.DataFrame: Sorted leaderboard with district statistics
     """
     leaderboard_service = await get_leaderboard_service()
-    return await leaderboard_service.build_district_leaderboard_last_week_ist(message_categories, processing_batch_size)
+    # Use week strategy explicitly - addressing review comment #5
+    week_strategy = TimeWindowFactory.create_strategy('week')
+    return await leaderboard_service.build_district_leaderboard(message_categories, processing_batch_size, week_strategy)
 
 async def build_district_leaderboard_with_strategy(
     strategy_type: str,
