+++ byoeb-v1/byoeb/byoeb/chat_app/run.py
@@ -50,7 +50,8 @@ async def lifespan(app: FastAPI):
         queue_producer_factory,
         text_translator
     )
-    from byoeb.apis.background_jobs import setup_scheduled_jobs, start_scheduler, stop_scheduler
+    from byoeb.apis.background_jobs import setup_scheduled_jobs
+    from byoeb.background_jobs.dependency_setup import start_scheduler, stop_scheduler
 
     await message_consumer.initialize()
     asyncio.create_task(message_consumer.listen())
