+++ byoeb-v1/byoeb/byoeb/services/databases/mongo_db/base.py
@@ -2,16 +2,40 @@ from byoeb_core.models.byoeb.user import User
 from byoeb_core.databases.mongo_db.base import BaseDocumentCollection
 from byoeb.factory import MongoDBFactory
 from byoeb_integrations.databases.mongo_db.azure.async_azure_cosmos_mongo_db import AsyncAzureCosmosMongoDBCollection
+from typing import Optional, TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from byoeb.repositories.repository_factory import RepositoryFactory
 
 class BaseMongoDBService:
-    """Base service class for MongoDB operations."""
+    """
+    Base service class for MongoDB operations.
+    Consolidates both direct MongoDB collection access and repository-based access.
+    Supports repository-based invocations while maintaining backward compatibility with direct collection access.
+    """
 
     def __init__(self, config, mongo_db_factory: MongoDBFactory):
         self._config = config
         self._mongo_db_factory = mongo_db_factory
+        self._repository_factory: Optional['RepositoryFactory'] = None  # Initialized lazily
 
     async def _get_collection_client(self, collection_name: str) -> BaseDocumentCollection:
-        """Get the MongoDB collection client based on the collection name."""
+        """
+        Get the MongoDB collection client based on the collection name.
+        Provides direct access to MongoDB collections for backward compatibility.
+        """
         mongo_db = await self._mongo_db_factory.get(self._config["app"]["db_provider"])
         collection = mongo_db.get_collection(collection_name)
         return AsyncAzureCosmosMongoDBCollection(collection=collection)
+
+    async def _get_repository_factory(self) -> 'RepositoryFactory':
+        """
+        Get or create repository factory instance.
+        Provides repository-based data access pattern support.
+        This method consolidates repository access functionality into BaseMongoDBService,
+        allowing all services to use repository-based invocations through a single class.
+        """
+        if self._repository_factory is None:
+            from byoeb.repositories.repository_factory import get_repository_factory as global_get_repository_factory
+            self._repository_factory = await global_get_repository_factory()
+        return self._repository_factory
