+++ byoeb-v1/byoeb/byoeb/services/message/__init__.py
@@ -1,4 +1,3 @@
-# Message services package
-from .message_service import MessageService
-
-__all__ = ['MessageService']
+# Message services package - MessageService has been removed
+# All functionality moved to byoeb.services.databases.mongo_db.MessageMongoDBService
+__all__ = []
