--- byoeb-v1/byoeb/byoeb/services/message/message_service.py
@@ -1,406 +1,0 @@
-"""
-Message Service for managing message-related operations.
-"""
-import pandas as pd
-from typing import List, Optional, Dict, Any
-from datetime import datetime, timezone, timedelta
-from collections import Counter, defaultdict
-from zoneinfo import ZoneInfo
-
-from byoeb.repositories.repository_factory import get_repository_factory
-from byoeb.services.user.user_service import UserService
-from byoeb.services.leaderboard.time_window_strategies import TimeWindowStrategy, TimeWindowFactory
-
-IST = ZoneInfo("Asia/Kolkata")
-
-class MessageService:
-    """Service class for message-related operations."""
-
-    def __init__(self, user_service: UserService):
-        self._user_service = user_service
-        self._repository_factory = None  # Initialized lazily
-
-    async def _get_repository_factory(self):
-        if self._repository_factory is None:
-            from byoeb.repositories.repository_factory import get_repository_factory as global_get_repository_factory
-            self._repository_factory = await global_get_repository_factory()
-        return self._repository_factory
-
-    def _district_of(self, user_obj) -> Optional[str]:
-        """
-        Extract district from user object.
-
-        Args:
-            user_obj: User object (can be SimpleNamespace or dict)
-
-        Returns:
-            Optional[str]: District name or None if not found/unknown
-        """
-        if not user_obj:
-            return None
-        loc = getattr(user_obj, "user_location", None) or {}
-        dist = loc.get("district") if hasattr(loc, "get") else getattr(loc, "district", None)
-        return str(dist).strip() if dist and str(dist).strip().lower() != "unknown" else None
-
-    async def build_district_leaderboard(
-        self, 
-        message_categories: Optional[List[str]] = None, 
-        processing_batch_size: int = 1000,
-        time_window_strategy: Optional[TimeWindowStrategy] = None
-    ) -> pd.DataFrame:
-        """
-        Builds a leaderboard of districts based on message activity for the specified time window.
-
-        Args:
-            message_categories: Optional list of message categories to filter by
-            processing_batch_size: Number of documents to process in each batch
-            time_window_strategy: Optional time window strategy (uses default if not provided)
-
-        Returns:
-            pd.DataFrame: Sorted leaderboard with district statistics
-        """
-        strategy = time_window_strategy or TimeWindowFactory.create_strategy('week')
-        start_timestamp, end_timestamp = strategy.calculate_window()
-
-        # Get repository instances
-        repository_factory = await self._get_repository_factory()
-        message_repository = await repository_factory.get_message_repository()
-
-        # Define projection for required fields only
-        required_fields_only = {"_id": 0, "message_data.user.user_id": 1, "message_data.incoming_timestamp": 1}
-
-        # Get messages using repository
-        message_documents = await message_repository.find_messages_by_time_range(
-            start_timestamp=start_timestamp,
-            end_timestamp=end_timestamp,
-            message_categories=message_categories,
-            projection=required_fields_only
-        )
-
-        # Sort messages by timestamp (descending)
-        message_documents.sort(key=lambda x: x.get("message_data", {}).get("incoming_timestamp", 0), reverse=True)
-
-        user_objects_cache = {}
-        district_message_counts = Counter()
-        district_unique_users = defaultdict(set)
-        district_first_message_timestamp = {}
-        district_last_message_timestamp = {}
-
-        # Process messages in batches
-        for i in range(0, len(message_documents), processing_batch_size):
-            message_batch = message_documents[i:i + processing_batch_size]
-
-            await self._user_service.hydrate_users(message_batch, user_objects_cache)
-
-            for message_document in message_batch:
-                message_data = message_document.get("message_data", {})
-                user_id = message_data.get("user", {}).get("user_id")
-                message_timestamp = message_data.get("incoming_timestamp")
-
-                if not isinstance(message_timestamp, int) or message_timestamp < start_timestamp or message_timestamp > end_timestamp:
-                    continue
-
-                user_object = user_objects_cache.get(user_id)
-                user_district = self._district_of(user_object)
-                if not user_district:
-                    continue
-
-                district_message_counts[user_district] += 1
-                if user_id:
-                    district_unique_users[user_district].add(user_id)
-
-                district_first_message_timestamp[user_district] = min(district_first_message_timestamp.get(user_district, message_timestamp), message_timestamp)
-                district_last_message_timestamp[user_district] = max(district_last_message_timestamp.get(user_district, message_timestamp), message_timestamp)
-
-        leaderboard_rows = [
-            {
-                "district": district_name,
-                "message_count": message_count,
-                "unique_users": len(district_unique_users[district_name]),
-                "first_seen": datetime.fromtimestamp(district_first_message_timestamp[district_name]).strftime("%d-%m-%Y %H:%M:%S"),
-                "last_seen": datetime.fromtimestamp(district_last_message_timestamp[district_name]).strftime("%d-%m-%Y %H:%M:%S")
-            }
-            for district_name, message_count in district_message_counts.items()
-        ]
-
-        if not leaderboard_rows:
-            return pd.DataFrame(
-                columns=["district", "message_count", "unique_users", "first_seen", "last_seen"]
-            )
-
-        return pd.DataFrame(leaderboard_rows).sort_values(by=["message_count", "unique_users"], ascending=False, ignore_index=True)
-
-    async def build_district_leaderboard_last_week_ist(
-        self, 
-        message_categories: Optional[List[str]] = None, 
-        processing_batch_size: int = 1000
-    ) -> pd.DataFrame:
-        """
-        Builds a leaderboard of districts based on message activity from the previous week in IST timezone.
-        This method is kept for backward compatibility.
-
-        Args:
-            message_categories: Optional list of message categories to filter by
-            processing_batch_size: Number of documents to process in each batch
-
-        Returns:
-            pd.DataFrame: Sorted leaderboard with district statistics
-        """
-        return await self.build_district_leaderboard(message_categories, processing_batch_size)
-
-    async def get_message_statistics_by_district(
-        self,
-        start_timestamp: int,
-        end_timestamp: int,
-        message_categories: Optional[List[str]] = None
-    ) -> Dict[str, Any]:
-        """
-        Get message statistics grouped by district for a specific time range.
-
-        Args:
-            start_timestamp: Start timestamp for the query
-            end_timestamp: End timestamp for the query
-            message_categories: Optional list of message categories to filter by
-
-        Returns:
-            Dict[str, Any]: Statistics including total messages, unique users, etc.
-        """
-        repository_factory = await self._get_repository_factory()
-        message_repository = await repository_factory.get_message_repository()
-
-        # Get messages using repository
-        message_documents = await message_repository.find_messages_by_time_range(
-            start_timestamp=start_timestamp,
-            end_timestamp=end_timestamp,
-            message_categories=message_categories
-        )
-
-        # Process statistics
-        total_messages = len(message_documents)
-        unique_users = set()
-        districts = set()
-
-        for message_document in message_documents:
-            message_data = message_document.get("message_data", {})
-            user_id = message_data.get("user", {}).get("user_id")
-            if user_id:
-                unique_users.add(user_id)
-
-        return {
-            "total_messages": total_messages,
-            "unique_users": len(unique_users),
-            "time_range": {
-                "start": datetime.fromtimestamp(start_timestamp).strftime("%Y-%m-%d %H:%M:%S"),
-                "end": datetime.fromtimestamp(end_timestamp).strftime("%Y-%m-%d %H:%M:%S")
-            },
-            "message_categories": message_categories or "all"
-        }
-
-    async def get_messages_by_user(
-        self,
-        user_id: str,
-        start_timestamp: Optional[int] = None,
-        end_timestamp: Optional[int] = None,
-        limit: int = 100
-    ) -> List[Dict[str, Any]]:
-        """
-        Get messages for a specific user within a time range.
-
-        Args:
-            user_id: The user ID to query
-            start_timestamp: Optional start timestamp
-            end_timestamp: Optional end timestamp
-            limit: Maximum number of messages to return
-
-        Returns:
-            List[Dict[str, Any]]: List of message documents
-        """
-        repository_factory = await self._get_repository_factory()
-        message_repository = await repository_factory.get_message_repository()
-
-        query = {"message_data.user.user_id": user_id}
-
-        if start_timestamp and end_timestamp:
-            query["message_data.incoming_timestamp"] = {
-                "$gte": start_timestamp,
-                "$lte": end_timestamp
-            }
-
-        return await message_repository.read_all(query, limit=limit)
-
-    async def get_message_count_by_category(
-        self,
-        start_timestamp: int,
-        end_timestamp: int
-    ) -> Dict[str, int]:
-        """
-        Get message count grouped by category for a specific time range.
-
-        Args:
-            start_timestamp: Start timestamp for the query
-            end_timestamp: End timestamp for the query
-
-        Returns:
-            Dict[str, int]: Category counts
-        """
-        repository_factory = await self._get_repository_factory()
-        message_repository = await repository_factory.get_message_repository()
-
-        # Get all messages in the time range
-        message_documents = await message_repository.find_messages_by_time_range(
-            start_timestamp=start_timestamp,
-            end_timestamp=end_timestamp
-        )
-
-        category_counts = Counter()
-        for message_document in message_documents:
-            message_data = message_document.get("message_data", {})
-            category = message_data.get("message_category", "unknown")
-            category_counts[category] += 1
-
-        return dict(category_counts)
-
-    async def send_bulk_messages(
-        self, 
-        phone_numbers: List[str], 
-        message_text: str,
-        debug_mode: bool = True,
-        test_mode: bool = False,
-        test_phone_number: str = "917567071072"
-    ) -> List[Dict[str, Any]]:
-        """
-        Send bulk messages to multiple phone numbers.
-
-        Args:
-            phone_numbers: List of phone numbers to send messages to
-            message_text: The text content to send
-            debug_mode: If True, prints detailed WhatsApp payloads instead of sending
-            test_mode: If True, only sends to test_phone_number instead of all numbers
-            test_phone_number: Phone number to use when test_mode is True
-
-        Returns:
-            List[Dict[str, Any]]: List of results for each message sent
-        """
-        from datetime import datetime
-
-        results = []
-
-        # If test_mode is enabled, only process the test phone number
-        target_phones = [test_phone_number] if test_mode else phone_numbers
-
-        for phone in target_phones:
-            message_payload = {
-                "object": "whatsapp_business_account",
-                "entry": [
-                    {
-                        "changes": [
-                            {
-                                "value": {
-                                    "messaging_product": "whatsapp",
-                                    "contacts": [{"wa_id": phone}],
-                                    "messages": [
-                                        {
-                                            "from": phone,
-                                            "id": f"custom-{phone}-{int(datetime.now().timestamp())}",
-                                            "timestamp": str(int(datetime.now().timestamp())),
-                                            "type": "text",
-                                            "text": {"body": message_text}
-                                        }
-                                    ]
-                                }
-                            }
-                        ]
-                    }
-                ]
-            }
-
-            if debug_mode:
-                # DEMO MODE: Print payload instead of sending
-                print("\n--- WhatsApp Message Payload ---")
-                print(f"To: {phone}")
-                print("Payload:", message_payload)
-                print("--- End Payload ---\n")
-
-                results.append({
-                    "phone": phone,
-                    "status": "debug_mode",
-                    "message": "Payload printed (not sent)",
-                    "payload": message_payload
-                })
-            else:
-                # ACTUAL SENDING MODE: Send real WhatsApp messages
-                try:
-                    from byoeb.chat_app.configuration.dependency_setup import channel_client_factory
-                    from byoeb_core.models.byoeb.message_context import ByoebMessageContext, MessageContext, ReplyContext, MessageTypes
-                    from byoeb_core.models.byoeb.user import User
-                    from datetime import datetime, timezone
-
-                    # Create a proper ByoebMessageContext for outgoing message
-                    byoeb_message = ByoebMessageContext(
-                        channel_type="whatsapp",
-                        message_category="leaderboard",
-                        user=User(
-                            user_id=f"leaderboard-{phone}",
-                            user_name="Leaderboard Bot",
-                            user_location={},
-                            user_language="en",
-                            user_type="bot",
-                            phone_number_id=phone,
-                            test_user=False,
-                            experts={},
-                            audience=[],
-                            created_timestamp=int(datetime.now(timezone.utc).timestamp()),
-                            activity_timestamp=int(datetime.now(timezone.utc).timestamp()),
-                            last_conversations=[],
-                            additional_info={}
-                        ),
-                        message_context=MessageContext(
-                            message_id=f"leaderboard-{phone}-{int(datetime.now().timestamp())}",
-                            message_type=MessageTypes.REGULAR_TEXT.value,
-                            message_source_text=message_text,
-                            message_english_text=message_text,
-                            media_info=None,
-                            additional_info={}
-                        ),
-                        reply_context=None,
-                        cross_conversation_id=None,
-                        cross_conversation_context=None,
-                        incoming_timestamp=int(datetime.now(timezone.utc).timestamp()),
-                        outgoing_timestamp=int(datetime.now(timezone.utc).timestamp())
-                    )
-
-                    # Get WhatsApp service and send the message
-                    from byoeb.services.channel.whatsapp import WhatsAppService
-                    whatsapp_service = WhatsAppService(channel_client_factory)
-                    requests = whatsapp_service.prepare_requests(byoeb_message)
-
-                    if requests:
-                        responses, message_ids = await whatsapp_service.send_requests(requests)
-                        print(f"✅ Sent to {phone} - Message ID: {message_ids[0] if message_ids else 'Unknown'}")
-                        results.append({
-                            "phone": phone,
-                            "status": "success",
-                            "message": f"Message sent successfully",
-                            "message_id": message_ids[0] if message_ids else None,
-                            "response": str(responses[0]) if responses else None
-                        })
-                    else:
-                        print(f"❌ No requests generated for {phone}")
-                        results.append({
-                            "phone": phone,
-                            "status": "error",
-                            "message": "No requests generated"
-                        })
-
-                except Exception as e:
-                    print(f"❌ Error sending to {phone}: {str(e)}")
-                    import traceback
-                    traceback.print_exc()
-                    results.append({
-                        "phone": phone,
-                        "status": "error",
-                        "message": f"Exception occurred: {str(e)}",
-                        "error": str(e)
-                    })
-
-        return results
