+++ byoeb-v1/byoeb/byoeb/services/user/__init__.py
@@ -1,6 +1,5 @@
 """
-User service module.
+User service module - UserService has been removed
+All functionality moved to byoeb.services.databases.mongo_db.UserMongoDBService
 """
-from .user_service import UserService
-
-__all__ = ["UserService"]
+__all__ = []
